+++ app/javascript/components/recordings/RoomRecordings.jsx
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 import {
-  Card, Stack, Container,
+  Card, Stack,
 } from 'react-bootstrap';
 import { useParams } from 'react-router-dom';
-import { Pagination } from 'semantic-ui-react';
 import useRoomRecordings from '../../hooks/queries/recordings/useRoomRecordings';
 import SearchBarQuery from '../shared/SearchBarQuery';
 import RecordingsList from './RecordingsList';
 import useRoomRecordingsProcessing from '../../hooks/queries/recordings/useRoomRecordingsProcessing';
 import RoomsRecordingRow from './RoomsRecordingRow';
+import Pagy from '../shared/Pagy';
 
 export default function RoomRecordings() {
   const [input, setInput] = useState('');
@@ -16,12 +16,6 @@ export default function RoomRecordings() {
   const { friendlyId } = useParams();
   const { isLoading, data: roomRecordings } = useRoomRecordings(friendlyId, input, page);
   const roomRecordingsProcessing = useRoomRecordingsProcessing(friendlyId);
-
-  const handlePage = (e, { activePage }) => {
-    const gotopage = { activePage };
-    const pagenum = gotopage.activePage;
-    setPage(pagenum);
-  };
 
   return (
     <div className="wide-background full-height-room">
@@ -39,17 +33,11 @@ export default function RoomRecordings() {
       </Card>
       {!isLoading
         && (
-        <Container className="text-center">
-          <Pagination
-            defaultActivePage={roomRecordings.meta.page}
-            totalPages={roomRecordings.meta.pages}
-            onPageChange={handlePage}
-            firstItem={null}
-            lastItem={null}
-            pointing
-            secondary
-          />
-        </Container>
+        <Pagy
+          page={roomRecordings.meta.page}
+          totalPages={roomRecordings.meta.pages}
+          setPage={setPage}
+        />
         )}
     </div>
   );
